requirements.txt
astroid==3.1.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dill==0.3.8
fonttools==4.51.0
iniconfig==2.0.0
isort==5.13.2
kiwisolver==1.4.5
matplotlib==3.8.4
mccabe==0.7.0
mpmath==1.3.0
mypy==1.10.0
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
pillow==10.3.0
platformdirs==4.2.1
pluggy==1.5.0
pylint==3.1.0
pyparsing==3.1.2
pytest==8.2.0
python-dateutil==2.9.0.post0
six==1.16.0
tomlkit==0.12.4
typing_extensions==4.11.0
